+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * `req.cookies` and `req.signedCookies` are now plain objects
+
 1.1.0 / 2014-05-12
 ==================